requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
black==21.7b0
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.8.3
requests==2.25.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
whitenoise==5.3.0
